refactoring:  rounding moved to a separate module
